+++ navigation-testing/src/commonMain/kotlin/com/freeletics/khonshu/navigation/deeplinks/DeepLinkDefinitions.kt
@@ -1,6 +1,5 @@
 package com.freeletics.khonshu.navigation.deeplinks
 
-import com.freeletics.khonshu.navigation.internal.InternalNavigationTestingApi
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.decodeFromString
 import net.peanuuutz.tomlkt.Toml
@@ -34,11 +33,6 @@ public value class PatternDefinition(
     init {
         // for validation purposes
         DeepLinkHandler.Pattern(value)
-    }
-
-    @OptIn(InternalNavigationTestingApi::class)
-    public fun replacePlaceholders(replacement: (String) -> String): String {
-        return DeepLinkHandler.Pattern(value).replacePlaceholders(replacement)
     }
 }